ProyectoCuencas/requirements.txt
streamlit
streamlit-folium
folium
geopandas
shapely
pandas
geopy
xlsxwriter
matplotlib.pyplot